+++ ethcore/Cargo.toml
@@ -7,14 +7,12 @@ version = "1.12.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 
 [dependencies]
-account-db = { path = "account-db" }
 account-state = { path = "account-state" }
 ansi_term = "0.11"
 basic-authority = { path = "./engines/basic-authority", optional = true} # used by test-helpers feature
 blooms-db = { path = "../util/blooms-db", optional = true }
 client-traits = { path = "./client-traits" }
 common-types = { path = "types" }
-crossbeam-utils = "0.6"
 engine = { path = "./engine" }
 env_logger = { version = "0.5", optional = true }
 ethabi = "8.0"
@@ -24,7 +22,6 @@ ethash = { path = "../ethash", optional = true }
 ethjson = { path = "../json", optional = true }
 ethkey = { path = "../accounts/ethkey", optional = true }
 ethcore-blockchain = { path = "./blockchain" }
-ethcore-bloom-journal = { path = "../util/bloom" }
 ethcore-call-contract = { path = "./call-contract" }
 ethcore-db = { path = "./db" }
 ethcore-io = { path = "../util/io" }
@@ -38,7 +35,6 @@ hash-db = "0.15.0"
 itertools = "0.5"
 journaldb = { path = "../util/journaldb" }
 keccak-hash = "0.2.0"
-keccak-hasher = { path = "../util/keccak-hasher" }
 kvdb = "0.1"
 kvdb-memorydb = { version = "0.1", optional = true }
 kvdb-rocksdb = { version = "0.1.3", optional = true }
@@ -47,9 +43,7 @@ log = "0.4"
 macros = { path = "../util/macros", optional = true }
 machine = { path = "./machine" }
 memory-cache = { path = "../util/memory-cache" }
-num_cpus = "1.2"
 parity-bytes = "0.1"
-parity-snappy = "0.1"
 parking_lot = "0.8"
 pod = { path = "pod", optional = true }
 trie-db = "0.15.0"
@@ -76,6 +70,7 @@ verification = { path = "./verification" }
 vm = { path = "vm" }
 
 [dev-dependencies]
+account-db = { path = "account-db" }
 blooms-db = { path = "../util/blooms-db" }
 ethcore-builtin = { path = "./builtin" }
 criterion = "0.2"
