+++ ethcore/snapshot/snapshot-tests/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "snapshot-tests"
+version = "0.1.0"
+authors = ["David Palm <dvdplm@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+accounts = { package = "ethcore-accounts", path = "../../../accounts" }
+account-db = { path = "../../account-db" }
+account-state = { path = "../../account-state" }
+blockchain = { package = "ethcore-blockchain", path = "../../blockchain" }
+bytes = { package = "parity-bytes", version = "0.1.0" }
+client-traits = { path = "../../client-traits" }
+common-types = { path = "../../types" }
+engine = { path = "../../engine", features = ["test-helpers"] }
+env_logger = "0.5"
+ethcore = { path = "../..", features = ["test-helpers"] }
+ethcore-db = { path = "../../db" }
+ethcore-io = { path = "../../../util/io" }
+ethereum-types = "0.6.0"
+ethkey = { path = "../../../accounts/ethkey" }
+ethtrie  = { package = "patricia-trie-ethereum", path = "../../../util/patricia-trie-ethereum" }
+hash-db = "0.15.0"
+journaldb = { path = "../../../util/journaldb" }
+keccak-hash = "0.2.0"
+keccak-hasher = { path = "../../../util/keccak-hasher" }
+kvdb = "0.1.0"
+kvdb-rocksdb = { version = "0.1.3" }
+log = "0.4.8"
+parking_lot = "0.8.0"
+rand = "0.6"
+rand_xorshift = "0.1.1"
+rlp = "0.4.2"
+snappy = { package = "parity-snappy", version ="0.1.0" }
+snapshot = { path = "../../snapshot", features = ["test-helpers"] }
+spec = { path = "../../spec" }
+tempdir = "0.3"
+trie-db = "0.15.0"
+trie-standardmap = "0.15.0"
+ethabi = "8.0"
+ethabi-contract = "8.0"
+ethabi-derive = "8.0"
+lazy_static = { version = "1.3" }
+triehash = { package = "triehash-ethereum", version = "0.2",  path = "../../../util/triehash-ethereum" }
