+++ ethcore/snapshot/snapshot-tests/src/proof_of_authority.rs
@@ -21,25 +21,25 @@ use std::sync::Arc;
 use std::str::FromStr;
 
 use accounts::AccountProvider;
-use ethcore::client::Client;
 use client_traits::{BlockChainClient, ChainInfo};
-use ethkey::Secret;
-use crate::tests::helpers as snapshot_helpers;
-use spec::Spec;
-use ethcore::test_helpers::generate_dummy_client_with_spec;
 use common_types::transaction::{Transaction, Action, SignedTransaction};
-use tempdir::TempDir;
-use log::trace;
-use ethereum_types::Address;
+use ethabi_contract::use_contract;
 use ethcore::{
-	test_helpers,
+	client::Client,
+	test_helpers::{self, generate_dummy_client_with_spec},
 	miner::{self, MinerService},
 };
+use ethereum_types::Address;
+use ethkey::Secret;
 use keccak_hash::keccak;
-use ethabi_contract::use_contract;
 use lazy_static::lazy_static;
+use log::trace;
+use spec::Spec;
+use tempdir::TempDir;
 
-use_contract!(test_validator_set, "../res/contracts/test_validator_set.json");
+use crate::helpers as snapshot_helpers;
+
+use_contract!(test_validator_set, "../../res/contracts/test_validator_set.json");
 
 const PASS: &'static str = "";
 const TRANSITION_BLOCK_1: usize = 2; // block at which the contract becomes activated.
