+++ ethcore/snapshot/snapshot-tests/src/test_validator_contract.json
