+++ ethcore/snapshot/src/account.rs
@@ -35,7 +35,7 @@ use rlp::{RlpStream, Rlp};
 use trie_db::{Trie, TrieMut};
 
 // An empty account -- these were replaced with RLP null data for a space optimization in v1.
-const ACC_EMPTY: BasicAccount = BasicAccount {
+pub const ACC_EMPTY: BasicAccount = BasicAccount {
 	nonce: U256([0, 0, 0, 0]),
 	balance: U256([0, 0, 0, 0]),
 	storage_root: KECCAK_NULL_RLP,
@@ -69,6 +69,7 @@ impl CodeState {
 	}
 }
 
+// todo[dvdplm] move to test helpers?
 // walk the account's storage trie, returning a vector of RLP items containing the
 // account address hash, account properties and the storage. Each item contains at most `max_storage_items`
 // storage records split according to snapshot format definition.
@@ -170,6 +171,7 @@ pub fn to_fat_rlps(
 	}
 }
 
+// todo[dvdplm] move to test helpers?
 // decode a fat rlp, and rebuild the storage trie as we go.
 // returns the account structure along with its newly recovered code,
 // if it exists.
@@ -243,184 +245,4 @@ pub fn from_fat_rlp(
 	};
 
 	Ok((acc, new_code))
-}
-
-#[cfg(test)]
-mod tests {
-	use account_db::{AccountDB, AccountDBMut};
-	use common_types::basic_account::BasicAccount;
-	use ethcore::test_helpers::get_temp_state_db;
-	use crate::tests::helpers::fill_storage;
-	use common_types::snapshot::Progress;
-
-	use keccak_hash::{KECCAK_EMPTY, KECCAK_NULL_RLP, keccak};
-	use ethereum_types::{H256, Address};
-	use hash_db::{HashDB, EMPTY_PREFIX};
-	use kvdb::DBValue;
-	use rlp::Rlp;
-
-	use std::collections::HashSet;
-
-	use super::{ACC_EMPTY, to_fat_rlps, from_fat_rlp};
-
-	#[test]
-	fn encoding_basic() {
-		let mut db = get_temp_state_db();
-		let addr = Address::random();
-
-		let account = BasicAccount {
-			nonce: 50.into(),
-			balance: 123456789.into(),
-			storage_root: KECCAK_NULL_RLP,
-			code_hash: KECCAK_EMPTY,
-			code_version: 0.into(),
-		};
-
-		let thin_rlp = ::rlp::encode(&account);
-		assert_eq!(::rlp::decode::<BasicAccount>(&thin_rlp).unwrap(), account);
-		let p = Progress::default();
-		let fat_rlps = to_fat_rlps(&keccak(&addr), &account, &AccountDB::from_hash(db.as_hash_db(), keccak(addr)), &mut Default::default(), usize::max_value(), usize::max_value(), &p).unwrap();
-		let fat_rlp = Rlp::new(&fat_rlps[0]).at(1).unwrap();
-		assert_eq!(from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr)), fat_rlp, H256::zero()).unwrap().0, account);
-	}
-
-	#[test]
-	fn encoding_version() {
-		let mut db = get_temp_state_db();
-		let addr = Address::random();
-
-		let account = BasicAccount {
-			nonce: 50.into(),
-			balance: 123456789.into(),
-			storage_root: KECCAK_NULL_RLP,
-			code_hash: KECCAK_EMPTY,
-			code_version: 1.into(),
-		};
-
-		let thin_rlp = ::rlp::encode(&account);
-		assert_eq!(::rlp::decode::<BasicAccount>(&thin_rlp).unwrap(), account);
-		let p = Progress::default();
-		let fat_rlps = to_fat_rlps(&keccak(&addr), &account, &AccountDB::from_hash(db.as_hash_db(), keccak(addr)), &mut Default::default(), usize::max_value(), usize::max_value(), &p).unwrap();
-		let fat_rlp = Rlp::new(&fat_rlps[0]).at(1).unwrap();
-		assert_eq!(from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr)), fat_rlp, H256::zero()).unwrap().0, account);
-	}
-
-	#[test]
-	fn encoding_storage() {
-		let mut db = get_temp_state_db();
-		let addr = Address::random();
-
-		let account = {
-			let acct_db = AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr));
-			let mut root = KECCAK_NULL_RLP;
-			fill_storage(acct_db, &mut root, &mut H256::zero());
-			BasicAccount {
-				nonce: 25.into(),
-				balance: 987654321.into(),
-				storage_root: root,
-				code_hash: KECCAK_EMPTY,
-				code_version: 0.into(),
-			}
-		};
-
-		let thin_rlp = ::rlp::encode(&account);
-		assert_eq!(::rlp::decode::<BasicAccount>(&thin_rlp).unwrap(), account);
-
-		let p = Progress::default();
-
-		let fat_rlp = to_fat_rlps(&keccak(&addr), &account, &AccountDB::from_hash(db.as_hash_db(), keccak(addr)), &mut Default::default(), usize::max_value(), usize::max_value(), &p).unwrap();
-		let fat_rlp = Rlp::new(&fat_rlp[0]).at(1).unwrap();
-		assert_eq!(from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr)), fat_rlp, H256::zero()).unwrap().0, account);
-	}
-
-	#[test]
-	fn encoding_storage_split() {
-		let mut db = get_temp_state_db();
-		let addr = Address::random();
-
-		let account = {
-			let acct_db = AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr));
-			let mut root = KECCAK_NULL_RLP;
-			fill_storage(acct_db, &mut root, &mut H256::zero());
-			BasicAccount {
-				nonce: 25.into(),
-				balance: 987654321.into(),
-				storage_root: root,
-				code_hash: KECCAK_EMPTY,
-				code_version: 0.into(),
-			}
-		};
-
-		let thin_rlp = ::rlp::encode(&account);
-		assert_eq!(::rlp::decode::<BasicAccount>(&thin_rlp).unwrap(), account);
-
-		let p = Progress::default();
-		let fat_rlps = to_fat_rlps(&keccak(addr), &account, &AccountDB::from_hash(db.as_hash_db(), keccak(addr)), &mut Default::default(), 500, 1000, &p).unwrap();
-		let mut root = KECCAK_NULL_RLP;
-		let mut restored_account = None;
-		for rlp in fat_rlps {
-			let fat_rlp = Rlp::new(&rlp).at(1).unwrap();
-			restored_account = Some(from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr)), fat_rlp, root).unwrap().0);
-			root = restored_account.as_ref().unwrap().storage_root.clone();
-		}
-		assert_eq!(restored_account, Some(account));
-	}
-
-	#[test]
-	fn encoding_code() {
-		let mut db = get_temp_state_db();
-
-		let addr1 = Address::random();
-		let addr2 = Address::random();
-
-		let code_hash = {
-			let mut acct_db = AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr1));
-			acct_db.insert(EMPTY_PREFIX, b"this is definitely code")
-		};
-
-		{
-			let mut acct_db = AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr2));
-			acct_db.emplace(code_hash.clone(), EMPTY_PREFIX, DBValue::from_slice(b"this is definitely code"));
-		}
-
-		let account1 = BasicAccount {
-			nonce: 50.into(),
-			balance: 123456789.into(),
-			storage_root: KECCAK_NULL_RLP,
-			code_hash,
-			code_version: 0.into(),
-		};
-
-		let account2 = BasicAccount {
-			nonce: 400.into(),
-			balance: 98765432123456789usize.into(),
-			storage_root: KECCAK_NULL_RLP,
-			code_hash,
-			code_version: 0.into(),
-		};
-
-		let mut used_code = HashSet::new();
-		let p1 = Progress::default();
-		let p2 = Progress::default();
-		let fat_rlp1 = to_fat_rlps(&keccak(&addr1), &account1, &AccountDB::from_hash(db.as_hash_db(), keccak(addr1)), &mut used_code, usize::max_value(), usize::max_value(), &p1).unwrap();
-		let fat_rlp2 = to_fat_rlps(&keccak(&addr2), &account2, &AccountDB::from_hash(db.as_hash_db(), keccak(addr2)), &mut used_code, usize::max_value(), usize::max_value(), &p2).unwrap();
-		assert_eq!(used_code.len(), 1);
-
-		let fat_rlp1 = Rlp::new(&fat_rlp1[0]).at(1).unwrap();
-		let fat_rlp2 = Rlp::new(&fat_rlp2[0]).at(1).unwrap();
-
-		let (acc, maybe_code) = from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr2)), fat_rlp2, H256::zero()).unwrap();
-		assert!(maybe_code.is_none());
-		assert_eq!(acc, account2);
-
-		let (acc, maybe_code) = from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(addr1)), fat_rlp1, H256::zero()).unwrap();
-		assert_eq!(maybe_code, Some(b"this is definitely code".to_vec()));
-		assert_eq!(acc, account1);
-	}
-
-	#[test]
-	fn encoding_empty_acc() {
-		let mut db = get_temp_state_db();
-		assert_eq!(from_fat_rlp(&mut AccountDBMut::from_hash(db.as_hash_db_mut(), keccak(Address::zero())), Rlp::new(&::rlp::NULL_RLP), H256::zero()).unwrap(), (ACC_EMPTY, None));
-	}
 }
