+++ ethcore/snapshot/src/io.rs
@@ -26,7 +26,6 @@ use std::fs::{self, File};
 use std::path::{Path, PathBuf};
 
 use bytes::Bytes;
-use client_traits::SnapshotWriter;
 use common_types::{
 	errors::{SnapshotError, EthcoreError},
 	snapshot::ManifestData,
@@ -36,7 +35,7 @@ use log::trace;
 use rlp::{RlpStream, Rlp};
 use rlp_derive::*;
 
-const SNAPSHOT_VERSION: u64 = 2;
+pub const SNAPSHOT_VERSION: u64 = 2;
 
 // (hash, len, offset)
 #[derive(RlpEncodable, RlpDecodable)]
@@ -184,6 +183,23 @@ pub trait SnapshotReader {
 	fn chunk(&self, hash: H256) -> io::Result<Bytes>;
 }
 
+
+/// Something which can write snapshots.
+/// Writing the same chunk multiple times will lead to implementation-defined
+/// behavior, and is not advised.
+pub trait SnapshotWriter {
+	/// Write a compressed state chunk.
+	fn write_state_chunk(&mut self, hash: H256, chunk: &[u8]) -> std::io::Result<()>;
+
+	/// Write a compressed block chunk.
+	fn write_block_chunk(&mut self, hash: H256, chunk: &[u8]) -> std::io::Result<()>;
+
+	/// Complete writing. The manifest's chunk lists must be consistent
+	/// with the chunks written.
+	fn finish(self, manifest: ManifestData) -> std::io::Result<()> where Self: Sized;
+}
+
+
 /// Packed snapshot reader.
 pub struct PackedReader {
 	file: File,
@@ -315,96 +331,5 @@ impl SnapshotReader for LooseReader {
 		let mut file = File::open(&path)?;
 		file.read_to_end(&mut buf)?;
 		Ok(buf)
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use tempdir::TempDir;
-	use keccak_hash::keccak;
-
-	use common_types::snapshot::ManifestData;
-	use super::{SnapshotWriter, SnapshotReader, PackedWriter, PackedReader, LooseWriter, LooseReader, SNAPSHOT_VERSION};
-
-	const STATE_CHUNKS: &'static [&'static [u8]] = &[b"dog", b"cat", b"hello world", b"hi", b"notarealchunk"];
-	const BLOCK_CHUNKS: &'static [&'static [u8]] = &[b"hello!", b"goodbye!", b"abcdefg", b"hijklmnop", b"qrstuvwxy", b"and", b"z"];
-
-	#[test]
-	fn packed_write_and_read() {
-		let tempdir = TempDir::new("").unwrap();
-		let path = tempdir.path().join("packed");
-		let mut writer = PackedWriter::new(&path).unwrap();
-
-		let mut state_hashes = Vec::new();
-		let mut block_hashes = Vec::new();
-
-		for chunk in STATE_CHUNKS {
-			let hash = keccak(&chunk);
-			state_hashes.push(hash.clone());
-			writer.write_state_chunk(hash, chunk).unwrap();
-		}
-
-		for chunk in BLOCK_CHUNKS {
-			let hash = keccak(&chunk);
-			block_hashes.push(hash.clone());
-			writer.write_block_chunk(keccak(&chunk), chunk).unwrap();
-		}
-
-		let manifest = ManifestData {
-			version: SNAPSHOT_VERSION,
-			state_hashes,
-			block_hashes,
-			state_root: keccak(b"notarealroot"),
-			block_number: 12345678987654321,
-			block_hash: keccak(b"notarealblock"),
-		};
-
-		writer.finish(manifest.clone()).unwrap();
-
-		let reader = PackedReader::new(&path).unwrap().unwrap();
-		assert_eq!(reader.manifest(), &manifest);
-
-		for hash in manifest.state_hashes.iter().chain(&manifest.block_hashes) {
-			reader.chunk(hash.clone()).unwrap();
-		}
-	}
-
-	#[test]
-	fn loose_write_and_read() {
-		let tempdir = TempDir::new("").unwrap();
-		let mut writer = LooseWriter::new(tempdir.path().into()).unwrap();
-
-		let mut state_hashes = Vec::new();
-		let mut block_hashes = Vec::new();
-
-		for chunk in STATE_CHUNKS {
-			let hash = keccak(&chunk);
-			state_hashes.push(hash.clone());
-			writer.write_state_chunk(hash, chunk).unwrap();
-		}
-
-		for chunk in BLOCK_CHUNKS {
-			let hash = keccak(&chunk);
-			block_hashes.push(hash.clone());
-			writer.write_block_chunk(keccak(&chunk), chunk).unwrap();
-		}
-
-		let manifest = ManifestData {
-			version: SNAPSHOT_VERSION,
-			state_hashes: state_hashes,
-			block_hashes: block_hashes,
-			state_root: keccak(b"notarealroot"),
-			block_number: 12345678987654321,
-			block_hash: keccak(b"notarealblock)"),
-		};
-
-		writer.finish(manifest.clone()).unwrap();
-
-		let reader = LooseReader::new(tempdir.path().into()).unwrap();
-		assert_eq!(reader.manifest(), &manifest);
-
-		for hash in manifest.state_hashes.iter().chain(&manifest.block_hashes) {
-			reader.chunk(hash.clone()).unwrap();
-		}
 	}
 }
