+++ ethcore/snapshot/src/watcher.rs
@@ -30,7 +30,7 @@ use log::{trace, warn};
 use parking_lot::Mutex;
 
 // helper trait for transforming hashes to numbers and checking if syncing.
-trait Oracle: Send + Sync {
+pub trait Oracle: Send + Sync {
 	fn to_number(&self, hash: H256) -> Option<u64>;
 
 	fn is_major_importing(&self) -> bool;
@@ -54,7 +54,7 @@ impl<F> Oracle for StandardOracle<F>
 }
 
 // helper trait for broadcasting a block to take a snapshot at.
-trait Broadcast: Send + Sync {
+pub trait Broadcast: Send + Sync {
 	fn take_at(&self, num: Option<u64>);
 }
 
@@ -98,6 +98,11 @@ impl Watcher {
 			history,
 		}
 	}
+
+	#[cfg(any(test, feature = "test-helpers"))]
+	pub fn new_test(oracle: Box<dyn Oracle>, broadcast: Box<dyn Broadcast>, period: u64, history: u64) -> Self {
+		Watcher { oracle, broadcast, period, history }
+	}
 }
 
 impl ChainNotify for Watcher {
@@ -120,81 +125,81 @@ impl ChainNotify for Watcher {
 	}
 }
 
-#[cfg(test)]
-mod tests {
-	use std::collections::HashMap;
-	use std::time::Duration;
-
-	use client_traits::ChainNotify;
-	use common_types::chain_notify::{NewBlocks, ChainRoute};
-
-	use ethereum_types::{H256, U256, BigEndianHash};
-
-	use super::{Broadcast, Oracle, Watcher};
-
-	struct TestOracle(HashMap<H256, u64>);
-
-	impl Oracle for TestOracle {
-		fn to_number(&self, hash: H256) -> Option<u64> {
-			self.0.get(&hash).cloned()
-		}
-
-		fn is_major_importing(&self) -> bool { false }
-	}
-
-	struct TestBroadcast(Option<u64>);
-	impl Broadcast for TestBroadcast {
-		fn take_at(&self, num: Option<u64>) {
-			if num != self.0 {
-				panic!("Watcher broadcast wrong number. Expected {:?}, found {:?}", self.0, num);
-			}
-		}
-	}
-
-	// helper harness for tests which expect a notification.
-	fn harness(numbers: Vec<u64>, period: u64, history: u64, expected: Option<u64>) {
-		const DURATION_ZERO: Duration = Duration::from_millis(0);
-
-		let hashes: Vec<_> = numbers.clone().into_iter().map(|x| BigEndianHash::from_uint(&U256::from(x))).collect();
-		let map = hashes.clone().into_iter().zip(numbers).collect();
-
-		let watcher = Watcher {
-			oracle: Box::new(TestOracle(map)),
-			broadcast: Box::new(TestBroadcast(expected)),
-			period,
-			history,
-		};
-
-		watcher.new_blocks(NewBlocks::new(
-			hashes,
-			vec![],
-			ChainRoute::default(),
-			vec![],
-			vec![],
-			DURATION_ZERO,
-			false
-		));
-	}
-
-	// helper
-
-	#[test]
-	fn should_not_fire() {
-		harness(vec![0], 5, 0, None);
-	}
-
-	#[test]
-	fn fires_once_for_two() {
-		harness(vec![14, 15], 10, 5, Some(10));
-	}
-
-	#[test]
-	fn finds_highest() {
-		harness(vec![15, 25], 10, 5, Some(20));
-	}
-
-	#[test]
-	fn doesnt_fire_before_history() {
-		harness(vec![10, 11], 10, 5, None);
-	}
-}
+//#[cfg(test)]
+//mod tests {
+//	use std::collections::HashMap;
+//	use std::time::Duration;
+//
+//	use client_traits::ChainNotify;
+//	use common_types::chain_notify::{NewBlocks, ChainRoute};
+//
+//	use ethereum_types::{H256, U256, BigEndianHash};
+//
+//	use super::{Broadcast, Oracle, Watcher};
+//
+//	struct TestOracle(HashMap<H256, u64>);
+//
+//	impl Oracle for TestOracle {
+//		fn to_number(&self, hash: H256) -> Option<u64> {
+//			self.0.get(&hash).cloned()
+//		}
+//
+//		fn is_major_importing(&self) -> bool { false }
+//	}
+//
+//	struct TestBroadcast(Option<u64>);
+//	impl Broadcast for TestBroadcast {
+//		fn take_at(&self, num: Option<u64>) {
+//			if num != self.0 {
+//				panic!("Watcher broadcast wrong number. Expected {:?}, found {:?}", self.0, num);
+//			}
+//		}
+//	}
+//
+//	// helper harness for tests which expect a notification.
+//	fn harness(numbers: Vec<u64>, period: u64, history: u64, expected: Option<u64>) {
+//		const DURATION_ZERO: Duration = Duration::from_millis(0);
+//
+//		let hashes: Vec<_> = numbers.clone().into_iter().map(|x| BigEndianHash::from_uint(&U256::from(x))).collect();
+//		let map = hashes.clone().into_iter().zip(numbers).collect();
+//
+//		let watcher = Watcher {
+//			oracle: Box::new(TestOracle(map)),
+//			broadcast: Box::new(TestBroadcast(expected)),
+//			period,
+//			history,
+//		};
+//
+//		watcher.new_blocks(NewBlocks::new(
+//			hashes,
+//			vec![],
+//			ChainRoute::default(),
+//			vec![],
+//			vec![],
+//			DURATION_ZERO,
+//			false
+//		));
+//	}
+//
+//	// helper
+//
+//	#[test]
+//	fn should_not_fire() {
+//		harness(vec![0], 5, 0, None);
+//	}
+//
+//	#[test]
+//	fn fires_once_for_two() {
+//		harness(vec![14, 15], 10, 5, Some(10));
+//	}
+//
+//	#[test]
+//	fn finds_highest() {
+//		harness(vec![15, 25], 10, 5, Some(20));
+//	}
+//
+//	#[test]
+//	fn doesnt_fire_before_history() {
+//		harness(vec![10, 11], 10, 5, None);
+//	}
+//}
